requirements.txt
flask==2.0.1
pillow==8.3.1
numpy==1.19.5
scipy==1.7.1
scikit-image==0.18.3
